+++ packages/expo-dev-launcher/ios/EXRCTAppDelegateInterceptor.mm
@@ -6,10 +6,6 @@
 #import <cxxreact/JSExecutor.h>
 #import <React/RCTCxxBridgeDelegate.h>
 
-#if __has_include(<RNReanimated/REAInitializer.h>)
-#import <RNReanimated/REAInitializer.h>
-#endif // __has_include(<RNReanimated/REAInitializer.h>)
-
 @interface DevClientAppDelegate () <RCTCxxBridgeDelegate>
 @end
 
@@ -55,18 +51,6 @@
 
 - (NSURL *)sourceURLForBridge:(RCTBridge *)bridge {
   return [self.interceptor sourceURLForBridge:bridge];
-}
-
-- (std::unique_ptr<facebook::react::JSExecutorFactory>)jsExecutorFactoryForBridge:(RCTBridge *)bridge
-{
-#if __has_include(<RNReanimated/REAInitializer.h>) \
-  && __has_include(<RNReanimated/UIResponder+Reanimated.h>) /* removed in react-native-reanimated@3.4.0 */ \
-  && !RCT_NEW_ARCH_ENABLED
-  // required and available only for react-native-reanimated < 3.4.0
-  reanimated::REAInitializer(bridge);
-#endif
-
-  return [super jsExecutorFactoryForBridge:bridge];
 }
 
 @end